Refactor: Extract parser into its own package and add registry
- Moved `model.go` to `model/model.go` to avoid circular dependencies.
- Created `parsers` package with a registry mechanism.
- Created `parsers/common` package for the default comment-based parser implementation.
- Registered the default parser as "comment-v1".
- Updated `gosubc` commands (`generate`, `list`, `validate`) to accept a `--parser-name` flag (default: "comment-v1").
- Updated imports and refactored tests to align with the new package structure.
- Fixed unused variable error in `issues_test.go`.

This change enables support for multiple parsing strategies in the future while maintaining backward compatibility via the default parser.
