+++ parsers/common/issue55_test.go
@@ -1,4 +1,4 @@
-package parser
+package common
 
 import (
 	"testing"
