+++ parsers/common/parser_regr_test.go
@@ -1,4 +1,4 @@
-package parser
+package common
 
 import (
 	"bytes"
