+++ parsers/common/parser_test.go
@@ -1,4 +1,4 @@
-package parser
+package common
 
 import (
 	"reflect"
