+++ parsers/common/sanitize.go
@@ -1,4 +1,4 @@
-package parser
+package common
 
 import (
 	"fmt"
