+++ parsers/common/sanitize_test.go
@@ -1,4 +1,4 @@
-package parser
+package common
 
 import (
 	"testing"
