+++ parsers/common/testdata/parser_regr/gofmt_flags.txtar
